+++ controllers/admin/AdminLocationsController.js
@@ -78,7 +78,7 @@ module.exports.deleteLocation = (req, res) => {
 
 module.exports.updateLocation = (req, res) => {
   Location.findById(req.params.id, async (err, location) => {
-    if (req.body.avatar && await fs.existsSync(path.resolve(process.env.IMAGE_UPLOAD_DIR, location.avatar))) {
+    if (req.body.avatar && location.avatar && await fs.existsSync(path.resolve(process.env.IMAGE_UPLOAD_DIR, location.avatar))) {
       await fs.unlinkSync(path.resolve(process.env.IMAGE_UPLOAD_DIR, location.avatar));
     }
     if (err) {
